1. Create a new Project for Prototype 2
Open Unity Hub and create an empty “Prototype 2” project in your course directory on the correct Unity version.
If you forget how to do this, refer to the instructions in Lesson 1.1 - Step 1
Click to download the Prototype 2 Starter Files, extract the compressed folder, and then import the .unitypackage into your project.
If you forget how to do this, refer to the instructions in Lesson 1.1 - Step 2
From the Project window, open the Prototype 2 scene and delete the SampleScene
In the top-right of the Unity Editor, change your Layout from Default to your custom layout